+++ src/components/choropleth2/legenda/hooks/use-municipal-legenda-data.ts
@@ -1,6 +1,6 @@
 import { municipalThresholds } from '~/components/choropleth2/municipal-thresholds';
 import { TMunicipalityMetricName } from '~/components/choropleth2/shared';
-import { useLegendaItems } from './useLegendaItems';
+import { useLegendaItems } from './use-legenda-items';
 
 export function useMunicipalLegendaData(metric: TMunicipalityMetricName) {
   const thresholdInfo = municipalThresholds[metric];
