+++ src/pages/gemeente/[code]/positief-geteste-mensen.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import Getest from '~/assets/test.svg';
 import { ChoroplethLegenda } from '~/components/choropleth2/legenda/ChoroplethLegenda';
-import { useMunicipalLegendaData } from '~/components/choropleth2/legenda/hooks/useMunicipalLegendaData';
+import { useMunicipalLegendaData } from '~/components/choropleth2/legenda/hooks/use-municipal-legenda-data';
 import { MunicipalityChoropleth } from '~/components/choropleth2/municipality-choropleth';
 import { createSelectMunicipalHandler } from '~/components/choropleth2/selectHandlers/createSelectMunicipalHandler';
 import { createPositiveTestedPeopleMunicipalTooltip } from '~/components/choropleth2/tooltips/municipal/createPositiveTestedPeopleMunicipalTooltip';
