+++ src/pages/landelijk/rioolwater.tsx
@@ -4,7 +4,7 @@ import { Spacer } from '~/components-styled/base';
 import { ChoroplethTile } from '~/components-styled/choropleth-tile';
 import { LineChartTile } from '~/components-styled/line-chart-tile';
 import { ValueAnnotation } from '~/components-styled/value-annotation';
-import { useSafetyRegionLegendaData } from '~/components/choropleth2/legenda/hooks/useSafetyRegionLegendaData';
+import { useSafetyRegionLegendaData } from '~/components/choropleth2/legenda/hooks/use-safety-region-legenda-data';
 import { SafetyRegionChoropleth } from '~/components/choropleth2/safety-region-choropleth';
 import { createSelectRegionHandler } from '~/components/choropleth2/selectHandlers/createSelectRegionHandler';
 import { createSewerRegionalTooltip } from '~/components/choropleth2/tooltips/region/createSewerRegionalTooltip';
