+++ src/pages/landelijk/verpleeghuis-besmette-locaties.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import Locatie from '~/assets/locaties.svg';
 import { LineChart } from '~/components/charts/index';
 import { ChoroplethLegenda } from '~/components/choropleth2/legenda/ChoroplethLegenda';
-import { useSafetyRegionLegendaData } from '~/components/choropleth2/legenda/hooks/useSafetyRegionLegendaData';
+import { useSafetyRegionLegendaData } from '~/components/choropleth2/legenda/hooks/use-safety-region-legenda-data';
 import { SafetyRegionChoropleth } from '~/components/choropleth2/safety-region-choropleth';
 import { createSelectRegionHandler } from '~/components/choropleth2/selectHandlers/createSelectRegionHandler';
 import { createInfectedLocationsRegionalTooltip } from '~/components/choropleth2/tooltips/region/createInfectedLocationsRegionalTooltip';
